third_party/blink/web_tests/external/wpt/css/css-transitions/properties-value-001-expected.txt

This is a testharness.js-based test.
Found 11 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] background-position length(pt) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(pc) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(px) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(em) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(ex) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(mm) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(cm) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position length(in) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] background-position percentage(%) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-position:2s" but got "background-position-x:2s"
[FAIL] word-spacing percentage(%) / values
  assert_not_equals: initial and target values may not match got disallowed value "0px"
[FAIL] word-spacing percentage(%) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "word-spacing:2s" but got ""
Harness: the test ran to completion.
